pyproject.toml
[tool.poetry]
name = "deeplearn"
version = "0.1.0"
description = ""
authors = ["intimyp <intimyp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = {version = "^2.1.1+cpu", source = "pytorch"}
numpy = "^1.26.2"
matplotlib = "^3.8.1"
plotly = "^5.18.0"
kaleido = "0.2.1"
hydra-core = "^1.3.2"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
download = "deeplearn.papers:download_papers"
train = "cli:train"

[tool.pyright]
venvPath = "."
venv = ".venv"
strict = ["deeplearn"]
